Move functions to service
